Satisfy Swift file length budget
