+++ Packages/CmuxMobileTerminal/Sources/CmuxMobileTerminal/TerminalHardwareKeyCommand.swift
@@ -1,0 +1,8 @@
+#if canImport(UIKit)
+import UIKit
+
+struct TerminalHardwareKeyCommand: Sendable {
+    let input: String
+    let modifierFlags: UIKeyModifierFlags
+}
+#endif
